requirements.txt
aiogram==3.*
python-dotenv==1.*
fastapi==0.*
uvicorn==0.*
aiosqlite==0.*
websockets==12.*


